Allow each floppy drive to be a different floppy_format.  Remove the use of the ARCEMFLOPPYFORM environment variable;  pick the format based on the image file size instead.  This means FDC.Sector and FDC.Sector_ReadAddr need to be set everytime FDC.CurrentDisc changes based on the format of the newly selected drive.  Unfortunately, the code `knows' the image format and makes use of that knowledge before it has been instructed to read any data from the disc.  Consequently, a DOS 720K image which has sectors numbered from 1 is initially accessed by RISC OS, before *it* knows it's a DOS disc, by sector 0.  This would cause an fseek(3) to -512 so it's checked for and 0 is used instead.
With that change, it's fairly straightforward to `eject' and `insert' a
floppy.  Once ejected the FDC 1772 emulation has fclose'd the image
file, although the OS, e.g. RO 3.11, seemingly caches the directory
contents.  Perhaps the FDC emulation needs to feed back a `disc ejected'
signal to the OS.  The X control pane has been altered is set to toggle
each drive's eject/insert status when 0, 1, 2, or 3 is pressed.  It
still uses the FloppyImage# file names, but the code's interface would
allow the user to pick an image file for insertion.  The X user
interface also tries to `insert' all four floppy drives on start-up to
allow booting from floppy;  this is how the old code used to work too.